Button for authorizing yourself with Google
Button sends you to googles oauth landing page and redirects you back to the github pages site. The redirect contains the auth token as a query parameter, or an error message if consent was not given. 
Authorization requests the `drive.file` scope which is limited to files created by this app.